server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gunicorn>=21.0.0
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=1.24.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
apscheduler>=3.10.4
tzdata>=2023.3
msal>=1.35.0
requests>=2.32.0
Office365-REST-Python-Client>=2.6.0
gspread>=6.0.0
google-auth-oauthlib>=1.2.0
groq>=0.11.0
jinja2>=3.1.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
prometheus-client>=0.18.0
schedule>=1.2.1
playwright>=1.42.0
email-validator
pydantic[email]
httpx
msal 
PyJWT[crypto]